feat: Promote tabs from labs to main package
