+++ modules/react/tabs/tabs/index.ts
