+++ modules/react/tabs/tabs/lib/cursor/index.ts
