+++ modules/react/tabs/tabs/spec/tsconfig.json
