+++ modules/react/tabs/tabs/stories/Tabs.stories.mdx
@@ -1,4 +1,4 @@
-import {Tabs} from '@workday/canvas-kit-labs-react/tabs';
+import {Tabs} from '@workday/canvas-kit-react/tabs';
 import {Specifications} from '@workday/canvas-kit-docs';
 
 import {Simple} from './examples/Simple';
