+++ modules/react/tabs/tabs/stories/examples/HoistedModel.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {space} from '@workday/canvas-kit-react/tokens';
 
 import {Button} from '@workday/canvas-kit-react/button';
-import {Tabs, useTabsModel} from '@workday/canvas-kit-labs-react/tabs';
+import {Tabs, useTabsModel} from '@workday/canvas-kit-react/tabs';
 
 export const HoistedModel = () => {
   const model = useTabsModel({
