+++ modules/react/tabs/tabs/stories/examples/Simple.tsx
@@ -1,20 +1,24 @@
 import React from 'react';
 import {space} from '@workday/canvas-kit-react/tokens';
 
-import {Tabs} from '@workday/canvas-kit-labs-react/tabs';
+import {Tabs} from '@workday/canvas-kit-react/tabs';
 
-export const DisabledTab = () => {
+export const Simple = () => {
   return (
     <Tabs>
       <Tabs.List>
         <Tabs.Item>First Tab</Tabs.Item>
-        <Tabs.Item disabled>Disabled Tab</Tabs.Item>
+        <Tabs.Item>Second Tab</Tabs.Item>
         <Tabs.Item>Third Tab</Tabs.Item>
+        <Tabs.Item>Fourth Tab</Tabs.Item>
+        <Tabs.Item>Fifth Tab</Tabs.Item>
       </Tabs.List>
       <div style={{marginTop: space.m}}>
         <Tabs.Panel>Contents of First Tab</Tabs.Panel>
-        <Tabs.Panel>Contents of Disabled Tab</Tabs.Panel>
+        <Tabs.Panel>Contents of Second Tab</Tabs.Panel>
         <Tabs.Panel>Contents of Third Tab</Tabs.Panel>
+        <Tabs.Panel>Contents of Fourth Tab</Tabs.Panel>
+        <Tabs.Panel>Contents of Fifth Tab</Tabs.Panel>
       </div>
     </Tabs>
   );
