+++ modules/react/tabs/tabs/stories/examples/SinglePanel.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {space} from '@workday/canvas-kit-react/tokens';
 
-import {Tabs} from '@workday/canvas-kit-labs-react/tabs';
+import {Tabs} from '@workday/canvas-kit-react/tabs';
 import {useTabsModel} from '../../lib/useTabsModel';
 
 export const SinglePanel = () => {
